Rename to allow support for other chip variants
- Update names on namespaces and classes.
- Rework enum.
- Rework sample.
