+++ devices/Bq2579x/29pins_package.png
