+++ devices/Bq2579x/Bq25798.nuspec
