+++ devices/Bq2579x/Properties/AssemblyInfo.cs
