+++ devices/Bq2579x/Settings.StyleCop
