+++ devices/Bq2579x/category.txt
