+++ devices/Bq2579x/packages.config
