+++ devices/Bq2579x/samples/Program.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Iot.Device.Bq25798;
+using Iot.Device.Bq2579x;
 using System.Device.I2c;
 using System.Diagnostics;
 using System.Threading;
@@ -13,12 +13,12 @@ using System.Threading;
 //Configuration.SetPinFunction(22, DeviceFunction.I2C1_CLOCK);
 //////////////////////////////////////////////////////////////////////
 
-I2cConnectionSettings settings = new(1, Bq25798.DefaultI2cAddress);
+I2cConnectionSettings settings = new(3, Bq2579x.DefaultI2cAddress);
 using I2cDevice device = I2cDevice.Create(settings);
-using Bq25798 charger = new(device);
+using Bq2579x charger = new(device);
 
 Debug.WriteLine("");
-Debug.WriteLine($"BQ25798 connected to I2C{device.ConnectionSettings.BusId}");
+Debug.WriteLine($"{(charger.Model == Model.Bq25792 ? "Bq25792" : "Bq25798")} connected to I2C{device.ConnectionSettings.BusId}");
 Debug.WriteLine("");
 
 Debug.WriteLine($"Minimum System Voltage is config @ {charger.MinimalSystemVoltage.VoltsDc:N3}V");
