+++ devices/Bq2579x/samples/Properties/AssemblyInfo.cs
