+++ devices/Bq2579x/samples/packages.config
