+++ devices/Bq2579x/version.json
